Opcodes: move storage type checks to cmdlib
So far, there was a static check whether or not a given
storage type of an opcode was enabled. This check is
no longer possibe at this layer, because the information
whether or not file/shared file storage is enabled must
be retrieved from the cluster's configuration. This patch
removes the static check (which was incomplete anyway)
and adds it to the business logic of cmdlib.

Signed-off-by: Helga Velroyen <helgav@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>