Initial version of cmdlib test framework
The initial version of the cmdlib test framework is able to execute LU's
with the following components mocked:

 * Configuration
 * IAllocator interface
 * Lock manager
 * MCPU processor
 * RPC runner

A base test class is provided which makes it easy to execute opcodes and
to perform asserts on the output.

Tests for the LUTestDelay logical unit demonstrate the functionality of
the test framework.

The framework is not yet fully complete, further work will be done as
the tests are written.

Signed-off-by: Thomas Thrainer <thomasth@google.com>
Reviewed-by: Michele Tartara <mtartara@google.com>