Remove test for mutable default parameter values
Opcode parameters contain default values which can be, for example, a
list, dict, or set. Because these values are mutable, 'Validate'
performs a deepcopy before constructing the opcode dict and,
therefore, changing these values will not affect the default value
stored in the opcode description. As a result, the default value can
be reused in several calls to 'Validate' without a problem. The test
to ensure immutable default values can thus be removed.

Signed-off-by: Jose A. Lopes <jabolopes@google.com>
Reviewed-by: Guido Trotter <ultrotter@google.com>