Add configure option on whether to install symlinks
As of Ganeti 2.10, all files are installed into a single, version
specific, directory. In this way, several Ganeti versions can be
installed in parallel. Appropriate symbolic links in the search path
will be added by the Ganeti upgrade procedure. For the initial
installation, however, links in the search path are needed. So make
the installation of symbolic links a configure-time option, defaulting
to the current behavior, that that installation is visible in the
path.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Thomas Thrainer <thomasth@google.com>