Fetch 'randomUuidFile' from 'ConstantUtils'
In 'Ganeti.Utils', update the reference to 'randomUuidFile' to fetch
the value from 'Ganeti.ConstantUtils' instead of the Python to Haskell
generated constant.  This allows breaking the dependency of
'Ganeti.Utils' on the constants module, which in turn will allow
constants to take their values from Haskell types without creating a
circular dependency. Update also the Python constant reference.

Signed-off-by: Jose A. Lopes <jabolopes@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>