Hs2Py constants: add constants from 'Ganeti.Types'
Add to Hs2Py constant generation all constants that are defined by the
Haskell defined declared in 'Ganeti.Types'.  This patch also fixes a
type inconsistency between Haskell and Python, namely, the constant
'cvAllEcodesStrings' which in Python was 'frozenset' and the type
generated in Haskell was 'List'.  With this patch, the Haskell type
becomes 'FrozenSet', the counterpart type to Python's 'frozenset'.

Signed-off-by: Jose A. Lopes <jabolopes@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>